# goasm: Implement amd64 optimised code