--- a/.github/workflows/build.yml
+++ b/.github/workflows/build.yml
@@ -29,5 +29,7 @@ jobs:
         uses: actions/setup-go@6edd4406fa81c3da01a34fa6f6343087c207a568 # v3.5.0
         with:
           go-version: ${{ matrix.go-version }}
+      - name: Run Verify
+        run: make verify
       - name: Run Cross Build
         run: make cross-build