--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,10 @@
 module github.com/pjbgf/sha1cd
 
-go 1.15
+go 1.19
+
+require (
+	github.com/mmcloughlin/avo v0.5.0 // indirect
+	golang.org/x/mod v0.6.0 // indirect
+	golang.org/x/sys v0.1.0 // indirect
+	golang.org/x/tools v0.2.0 // indirect
+)