--- a/sha1cdblock_noasm.go
+++ b/sha1cdblock_noasm.go
@@ -1,0 +1,8 @@
+//go:build !amd64 || noasm || !gc
+// +build !amd64 noasm !gc
+
+package sha1cd
+
+func block(dig *digest, p []byte) {
+	blockGeneric(dig, p)
+}