--- a/test/testdata/fuzz/Fuzz_DeviationDetection/6cfa486437287a4340595f43c88ee06028ccbbf64661768e9596789910b0b2c1
+++ b/test/testdata/fuzz/Fuzz_DeviationDetection/6cfa486437287a4340595f43c88ee06028ccbbf64661768e9596789910b0b2c1
@@ -1,2 +1,0 @@
-go test fuzz v1
-[]byte("0009cA0$")