… go.sum
… sha1cdblock_amd64.s